+++ client/src/components/game/GameDashboard.jsx
@@ -21,7 +21,7 @@ export default function GameDashboard() {
   useEffect(() => {
     const interval = setInterval(() => {
       // console.log(petState);
-      gameTick(1000, petState);
+      gameTick(petState);
     }, 1000);
 
     return () => clearInterval(interval);
