+++ client/src/lib/petStatus.js
@@ -1,9 +1,9 @@
 export function getRating({ energy, food, happiness }) {
   //Status stuff
   let status = RATING.HAPPY;
-  if (food <= 15 && food < happiness && food < energy) status = RATING.HUNGRY;
-  else if (happiness <= 15 && happiness < energy) status = RATING.SAD;
-  else if (energy <= 15) status = RATING.TIRED;
+  if (food <= 15 && food < happiness && food < energy) return RATING.HUNGRY;
+  else if (happiness <= 15 && happiness < energy) return RATING.SAD;
+  else if (energy <= 15) return RATING.TIRED;
   if (energy <= 15 && happiness < 30) status = RATING.UPSET;
   if (energy - happiness > 20) status = RATING.UPSET;
   if (happiness - food > 20 && food < 20) status = RATING.HUNGRY;
@@ -23,4 +23,5 @@ export const STATUS = {
   PLAYING: 'Playing',
   SLEEPING: 'Sleeping',
   IDLE: 'Idle',
+  DEAD: 'Dead',
 };
